Dynamic solution for finding tables implemented
