+++ Assets/Scripts/AI.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 7460b58a4135cba4d920ae7ff3b7ea93
+guid: 3eecd3f4d10b1414380f738ba441d362
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
